Implement better Screen Zoom levels
This extends the density scaling options and mechanics
in the Screen Zoom section.

It gets extended from 3 options to 5 options.

The options for scaling after this patch are as follows:
smallest, small, default, large, largest.
(Updated translations will be needed!)

MIN_SCALE and MAX_SCALE are adjusted to unlock
the options we want and to still match
the formula.

Also MIN_DIMENSION_DP could be edited safely
for the same reason.

The formula for the density scaling is the following:

DisplayMetrics.DENSITY_MEDIUM * minDimensionPx / MIN_DIMENSION_DP;

DisplayMetrics.DENSITY_MEDIUM = 160
minDimensionPx = 1080

After these changes, DisplayMetrics will take care
of properly applying the requested changes to the system.

We are also removing support for an extremly large
size, since it would make the device unusable.

This resolves NextBit Ticket PARANOID-13.

Change-Id: I55c13e1bd74fede0db0bd79cd501e1a4f56c1f72
Signed-off-by: Alex Naidis <alex.naidis@linux.com>

Conflicts:
	packages/SettingsLib/src/com/android/settingslib/display/DisplayDensityUtils.java