Introduce Keypress Boost
Common cases of jank are button and key presses.
So it is useful to introduce a boost which intercepts key
presses and applies fine tuned boost patterns based on the key/button
type.

A boost pattern consists of a boost duration and device/chipset
specific boost params.

Keypress Boost allows specifying a weak and a strong boost pattern.
The weak boost pattern is applied to common button presses.
Soft buttons and hardware buttons are handled in this implementation.
Additionally we send a weak keypress boost for the media play event
to minimize the media startup latency and for the "enter" event from
a keyboard.

A strong Keypress Boost is dispatched for the power button press and
for the camera.

The defined boost patterns for Keypress Boost can be used for other
boosting features which would be dependant on Keypress Boost.

Change-Id: Ie192475a8fa5cbf43b0cd9913f6207bb9c3c0d3d
Signed-off-by: Alex Naidis <alex.naidis@linux.com>

Conflicts:
	services/core/java/com/android/server/policy/PhoneWindowManager.java